Rename NotifierBase to ReporterBase
