+++ master/docs/manual/configuration/reporters/index.rst
@@ -9,7 +9,7 @@ Reporters
     :hidden:
     :maxdepth: 2
 
-    notifier_base
+    reporter_base
     bitbucket_server_core_api_status
     bitbucket_server_pr_comment_push
     bitbucket_server_status
@@ -72,4 +72,4 @@ The following reporters are available:
  * :bb:reporter:`TelegramBot`
  * :bb:reporter:`ZulipStatusPush`
 
-Most of the report generators derive from :class:`NotifierBase` which implements basic reporter management functionality.
+Most of the report generators derive from :class:`ReporterBase` which implements basic reporter management functionality.
